+++ kube-controllers/pkg/controllers/networkpolicy/policy_controller_fv_test.go
@@ -1,0 +1,275 @@
+// Copyright (c) 2017-2022 Tigera, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package networkpolicy_test
+
+import (
+	"context"
+	"io/ioutil"
+	"os"
+	"time"
+
+	networkingv1 "k8s.io/api/networking/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	api "github.com/projectcalico/api/pkg/apis/projectcalico/v3"
+
+	"github.com/projectcalico/calico/felix/fv/containers"
+	"github.com/projectcalico/calico/kube-controllers/tests/testutils"
+	"github.com/projectcalico/calico/libcalico-go/lib/apiconfig"
+	client "github.com/projectcalico/calico/libcalico-go/lib/clientv3"
+	"github.com/projectcalico/calico/libcalico-go/lib/options"
+)
+
+var _ = Describe("Calico networkpolicy controller FV tests (etcd mode)", func() {
+	var (
+		etcd              *containers.Container
+		policyController  *containers.Container
+		apiserver         *containers.Container
+		calicoClient      client.Interface
+		k8sClient         *kubernetes.Clientset
+		controllerManager *containers.Container
+	)
+
+	BeforeEach(func() {
+		// Run etcd.
+		etcd = testutils.RunEtcd()
+		calicoClient = testutils.GetCalicoClient(apiconfig.EtcdV3, etcd.IP, "")
+
+		// Run apiserver.
+		apiserver = testutils.RunK8sApiserver(etcd.IP)
+
+		// Write out a kubeconfig file
+		kconfigfile, err := ioutil.TempFile("", "ginkgo-policycontroller")
+		Expect(err).NotTo(HaveOccurred())
+		defer os.Remove(kconfigfile.Name())
+		data := testutils.BuildKubeconfig(apiserver.IP)
+		_, err = kconfigfile.Write([]byte(data))
+		Expect(err).NotTo(HaveOccurred())
+
+		// Make the kubeconfig readable by the container.
+		Expect(kconfigfile.Chmod(os.ModePerm)).NotTo(HaveOccurred())
+
+		// Run the controller.
+		policyController = testutils.RunPolicyController(apiconfig.EtcdV3, etcd.IP, kconfigfile.Name(), "")
+
+		k8sClient, err = testutils.GetK8sClient(kconfigfile.Name())
+		Expect(err).NotTo(HaveOccurred())
+
+		// Wait for the apiserver to be available.
+		Eventually(func() error {
+			_, err := k8sClient.CoreV1().Namespaces().List(context.Background(), metav1.ListOptions{})
+			return err
+		}, 30*time.Second, 1*time.Second).Should(BeNil())
+
+		// Run controller manager.  Empirically it can take around 10s until the
+		// controller manager is ready to create default service accounts, even
+		// when the k8s image has already been downloaded to run the API
+		// server.  We use Eventually to allow for possible delay when doing
+		// initial pod creation below.
+		controllerManager = testutils.RunK8sControllerManager(apiserver.IP)
+	})
+
+	AfterEach(func() {
+		controllerManager.Stop()
+		policyController.Stop()
+		apiserver.Stop()
+		etcd.Stop()
+	})
+
+	Context("NetworkPolicy FV tests", func() {
+		var (
+			policyName        string
+			genPolicyName     string
+			policyNamespace   string
+			policyLabels      map[string]string
+			policyAnnotations map[string]string
+		)
+
+		BeforeEach(func() {
+			// Create a Kubernetes NetworkPolicy.
+			policyName = "jelly"
+			genPolicyName = "knp.default." + policyName
+			policyNamespace = "default"
+			policyAnnotations = map[string]string{
+				"annotK": "annotV",
+			}
+			policyLabels = map[string]string{
+				"labelK": "labelV",
+			}
+
+			np := &networkingv1.NetworkPolicy{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:        policyName,
+					Namespace:   policyNamespace,
+					Annotations: policyAnnotations,
+					Labels:      policyLabels,
+				},
+				Spec: networkingv1.NetworkPolicySpec{
+					PodSelector: metav1.LabelSelector{
+						MatchLabels: map[string]string{
+							"fools": "gold",
+						},
+					},
+				},
+			}
+
+			// Create the NP.
+			Eventually(func() error {
+				_, err := k8sClient.NetworkingV1().NetworkPolicies(policyNamespace).Create(context.Background(),
+					np, metav1.CreateOptions{})
+				return err
+			}, time.Second*5).ShouldNot(HaveOccurred())
+
+			// Wait for it to appear in Calico's etcd.
+			Eventually(func() *api.NetworkPolicy {
+				policy, _ := calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+				return policy
+			}, time.Second*5, 500*time.Millisecond).ShouldNot(BeNil())
+		})
+
+		It("should re-write policies in etcd when deleted in order to match policies in k8s", func() {
+			// Delete the Policy.
+			_, err := calicoClient.NetworkPolicies().Delete(context.Background(), policyNamespace, genPolicyName, options.DeleteOptions{})
+			Expect(err).ShouldNot(HaveOccurred())
+
+			// Wait for the policy-controller to write it back.
+			Eventually(func() error {
+				_, err := calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+				return err
+			}, time.Second*15, 500*time.Millisecond).ShouldNot(HaveOccurred())
+		})
+
+		It("should re-program policies that have changed in etcd", func() {
+			p, err := calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+			By("getting the policy", func() {
+				Expect(err).ShouldNot(HaveOccurred())
+			})
+
+			By("updating the selector on the policy", func() {
+				p.Spec.Selector = "ping == 'pong'"
+				p2, err := calicoClient.NetworkPolicies().Update(context.Background(), p, options.SetOptions{})
+				Expect(err).ShouldNot(HaveOccurred())
+				Expect(p2.Spec.Selector).To(Equal("ping == 'pong'"))
+			})
+
+			By("waiting for the controller to write back the correct selector", func() {
+				Eventually(func() string {
+					p, _ := calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+					return p.Spec.Selector
+				}, time.Second*15, 500*time.Millisecond).Should(Equal("projectcalico.org/orchestrator == 'k8s' && fools == 'gold'"))
+			})
+		})
+
+		It("should delete policies when they are deleted from the Kubernetes API", func() {
+			By("deleting the policy", func() {
+				err := k8sClient.NetworkingV1().NetworkPolicies(policyNamespace).Delete(context.Background(),
+					policyName, metav1.DeleteOptions{})
+				Expect(err).NotTo(HaveOccurred())
+			})
+
+			By("waiting for it to be removed from etcd", func() {
+				Eventually(func() error {
+					_, err := calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+					return err
+				}, time.Second*15, 500*time.Millisecond).Should(HaveOccurred())
+			})
+		})
+	})
+
+	Context("NetworkPolicy egress FV tests", func() {
+		var (
+			policyName      string
+			genPolicyName   string
+			policyNamespace string
+		)
+
+		BeforeEach(func() {
+			// Create a Kubernetes NetworkPolicy.
+			policyName = "jelly"
+			genPolicyName = "knp.default." + policyName
+			policyNamespace = "default"
+			np := &networkingv1.NetworkPolicy{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      policyName,
+					Namespace: policyNamespace,
+				},
+				Spec: networkingv1.NetworkPolicySpec{
+					PodSelector: metav1.LabelSelector{
+						MatchLabels: map[string]string{
+							"fools": "gold",
+						},
+					},
+					Egress: []networkingv1.NetworkPolicyEgressRule{
+						{
+							To: []networkingv1.NetworkPolicyPeer{
+								{
+									IPBlock: &networkingv1.IPBlock{
+										CIDR:   "192.168.0.0/16",
+										Except: []string{"192.168.3.0/24"},
+									},
+								},
+							},
+						},
+					},
+					PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeEgress},
+				},
+			}
+
+			// Create the NP.
+			Eventually(func() error {
+				_, err := k8sClient.NetworkingV1().NetworkPolicies(policyNamespace).Create(context.Background(),
+					np, metav1.CreateOptions{})
+				return err
+			}, time.Second*5).ShouldNot(HaveOccurred())
+
+			// Wait for it to appear in Calico's etcd.
+			Eventually(func() *api.NetworkPolicy {
+				p, _ := calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+				return p
+			}, time.Second*15, 500*time.Millisecond).ShouldNot(BeNil())
+		})
+
+		It("should contain the correct rules", func() {
+			var p *api.NetworkPolicy
+			By("getting the network policy created by the controller", func() {
+				Eventually(func() error {
+					var err error
+					p, err = calicoClient.NetworkPolicies().Get(context.Background(), policyNamespace, genPolicyName, options.GetOptions{})
+					return err
+				}, time.Second*10, 500*time.Millisecond).Should(BeNil())
+			})
+
+			By("checking the policy's selector is correct", func() {
+				Expect(p.Spec.Selector).Should(Equal("projectcalico.org/orchestrator == 'k8s' && fools == 'gold'"))
+			})
+
+			By("checking the policy's egress rule is correct", func() {
+				Expect(len(p.Spec.Egress)).Should(Equal(1))
+			})
+
+			By("checking the policy has type 'Egress'", func() {
+				Expect(p.Spec.Types).Should(Equal([]api.PolicyType{api.PolicyTypeEgress}))
+			})
+
+			By("checking the policy has no ingress rule", func() {
+				Expect(len(p.Spec.Ingress)).Should(Equal(0))
+			})
+		})
+	})
+})
