+++ kube-controllers/pkg/controllers/pod/pod_suite_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2018 Tigera, Inc. All rights reserved.
+// Copyright (c) 2021 Tigera, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,27 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package fv_test
+package pod_test
 
 import (
+	"testing"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
 
 	"github.com/projectcalico/calico/libcalico-go/lib/testutils"
 
-	"testing"
-
 	"github.com/onsi/ginkgo/reporters"
 )
 
 func init() {
 	testutils.HookLogrusForGinkgo()
-	logrus.SetLevel(logrus.InfoLevel)
+	logrus.SetLevel(logrus.DebugLevel)
 }
 
-func TestFv(t *testing.T) {
+func TestConverter(t *testing.T) {
 	RegisterFailHandler(Fail)
-	junitReporter := reporters.NewJUnitReporter("../../report/fv_suite.xml")
-	RunSpecsWithDefaultAndCustomReporters(t, "Fv Suite", []Reporter{junitReporter})
+	junitReporter := reporters.NewJUnitReporter("../../report/pod_controller_suite.xml")
+	RunSpecsWithDefaultAndCustomReporters(t, "Pod controller suite", []Reporter{junitReporter})
 }
